// forte-admin/src/lib/types/common.ts
export interface File {
  id: number;

  key: string;

  fileUrl: string;

  filename: string;

  createdAt?: string;

  slug?: string;
}

export interface User {
  id: number;

  firstName: string;

  lastName: string;
}

export type StatusTypes = "active" | "inactive";

export type RolesTypes = "provider" | "admin" | "user";

export enum SortValues {
  ASC = "asc",
  DESC = "desc",
}